move translator to utils folder, so that it is ok to be used from a component
