+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Datagrid/Datagrid.js
@@ -5,7 +5,6 @@ import React from 'react';
 import equal from 'fast-deep-equal';
 import InfiniteScroller from '../../components/InfiniteScroller';
 import Pagination from '../../components/Pagination';
-import {translate} from '../../services/Translator';
 import DatagridStore from './stores/DatagridStore';
 import datagridAdapterRegistry from './registries/DatagridAdapterRegistry';
 import AbstractAdapter from './adapters/AbstractAdapter';
