+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/index.js
@@ -1,10 +1,8 @@
 // @flow
 import {bundleReady} from './Bundles';
-import {translate} from './Translator';
 import ResourceRequester from './ResourceRequester';
 
 export {
     bundleReady,
-    translate,
     ResourceRequester,
 };
