+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/Translator/README.md
