+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/Translator/Translator.js
