+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/Translator/index.js
