+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/Translator/tests/Translator.test.js
