+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/Translator/types.js
