+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/index.js
@@ -1,0 +1,6 @@
+// @flow
+import {translate} from './Translator';
+
+export {
+    translate,
+};
