+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/Datagrid/tests/adapters/MediaCardOverviewAdapter.test.js
@@ -3,7 +3,7 @@ import {render} from 'enzyme';
 import React from 'react';
 import MediaCardOverviewAdapter from '../../adapters/MediaCardOverviewAdapter';
 
-jest.mock('sulu-admin-bundle/services', () => ({
+jest.mock('sulu-admin-bundle/utils', () => ({
     translate: function(key) {
         switch (key) {
             case 'sulu_media.copy_url':
