+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaCollection/tests/MediaCollection.test.js
@@ -88,7 +88,7 @@ jest.mock('../../../stores/CollectionStore', () => {
     return jest.fn();
 });
 
-jest.mock('sulu-admin-bundle/services', () => ({
+jest.mock('sulu-admin-bundle/utils', () => ({
     translate: function(key) {
         switch (key) {
             case 'sulu_media.all_media':
@@ -109,7 +109,7 @@ jest.mock('sulu-admin-bundle/services', () => ({
     },
 }));
 
-jest.mock('sulu-admin-bundle/services/Translator', () => ({
+jest.mock('sulu-admin-bundle/utils/Translator', () => ({
     translate: function(key) {
         switch (key) {
             case 'sulu_admin.page':
