+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelection/tests/MediaSelection.test.js
@@ -93,7 +93,7 @@ jest.mock('../../../stores/CollectionStore', () => jest.fn(function() {
     this.destroy = jest.fn();
 }));
 
-jest.mock('sulu-admin-bundle/services', () => ({
+jest.mock('sulu-admin-bundle/utils', () => ({
     translate: function(key) {
         switch (key) {
             case 'sulu_media.all_media':
@@ -124,7 +124,7 @@ jest.mock('sulu-admin-bundle/services', () => ({
     },
 }));
 
-jest.mock('sulu-admin-bundle/services/Translator', () => ({
+jest.mock('sulu-admin-bundle/utils/Translator', () => ({
     translate: function(key) {
         switch (key) {
             case 'sulu_admin.page':
